reorganized file structure & moved views from components into the views folder
