+++ src/components/clearHistoryModal/index.js
