+++ src/components/index.js
@@ -1,3 +1,2 @@
 export { default as Header } from './header'
 export { default as Footer } from './footer'
-export { default as History } from './history'
