+++ src/components/resetStorageModal/index.js
