+++ src/components/rows/index.js
@@ -1,0 +1,3 @@
+export { default as Row } from './row.js'
+export { default as Heading } from './heading.js'
+export { default as SectionWrapper } from './sectionWrapper'
